requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.7
djangorestframework==3.10.3
idna==2.8
numpy==1.17.4
psycopg2==2.8.4
pytz==2019.3
requests==2.22.0
scipy==1.3.2
sqlparse==0.3.0
urllib3==1.25.7
